requirements.txt
dagster==1.9
# Do we even need these here and not in the pyproject.toml?
dagster-dbt
dagster-aws

dbt-core==1.8
dbt-postgres
sqlalchemy
psycopg2-binary
python-dotenv
boto3
pandas

sshtunnel
paramiko

black
pylint
mypy
pytest